pyproject.toml
[project]
name = "ds-rpc-01"
version = "0.1.0"
description = "Starter project for the RPC-01: Internal Chatbot with Role Based Access Control"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "chromadb>=0.5.5",
    "sentence-transformers>=2.6.1",
    "groq>=0.11.0",
    "uvicorn[standard]>=0.23.0",
    "streamlit>=1.12.0,<2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "duckdb>=0.10.2",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
